+++ tmp/templates/app_template/config/boot.rb
@@ -1,1 +1,1 @@
-require "rails/all"
+require "bundler/setup" # Set up gems listed in the Gemfile.